# Conflicts:
#	src/main/java/legend/game/submap/SMap.java

+++ src/main/java/legend/game/types/Model124.java
@@ -165,8 +165,10 @@ public class Model124 {
   }
 
   public void deleteModelParts() {
-    for(final ModelPart10 part : this.modelParts_00) {
-      part.delete();
+    if(this.modelParts_00 != null) {
+      for(final ModelPart10 part : this.modelParts_00) {
+        part.delete();
+      }
     }
   }
 }
